requirements.txt
aiofiles==23.2.1
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
asyncer==0.0.2
attrs==23.1.0
backoff==2.2.1
beautifulsoup4==4.12.2
benzinga==1.21
bidict==0.22.1
certifi==2023.7.22
chainlit==0.7.301
charset-normalizer==3.3.0
click==8.1.7
dataclasses-json==0.5.14
Deprecated==1.2.14
diskcache==5.6.3
dnspython==2.4.2
exceptiongroup==1.1.3
fastapi==0.100.1
fastapi-socketio==0.0.10
filetype==1.2.0
FLAML==2.1.1
frozenlist==1.4.0
googleapis-common-protos==1.61.0
greenlet==3.0.0
grpcio==1.59.0
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.315
langsmith==0.0.44
Lazify==0.4.0
loguru==0.7.2
MarkupSafe==2.1.3
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.8
nodeenv==1.8.0
numpy==1.26.1
openai==0.28.1
opentelemetry-api==1.20.0
opentelemetry-exporter-otlp==1.20.0
opentelemetry-exporter-otlp-proto-common==1.20.0
opentelemetry-exporter-otlp-proto-grpc==1.20.0
opentelemetry-exporter-otlp-proto-http==1.20.0
opentelemetry-instrumentation==0.41b0
opentelemetry-proto==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-semantic-conventions==0.41b0
packaging==23.2
pinecone-client==2.2.4
prisma==0.10.0
protobuf==4.24.4
pyautogen==0.1.11
pydantic==2.4.2
pydantic_core==2.10.1
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.8.0
python-graphql-client==0.4.3
python-multipart==0.0.6
python-socketio==5.10.0
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.22
starlette==0.27.0
structlog==23.2.0
syncer==2.0.3
tenacity==8.2.3
termcolor==2.3.0
tiktoken==0.5.1
tomli==2.0.1
tomlkit==0.12.1
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.8.0
uptrace==1.20.2
urllib3==1.25.10
uvicorn==0.23.2
watchfiles==0.20.0
websockets==12.0
wrapt==1.15.0
wsproto==1.2.0
yarl==1.9.2
zipp==3.17.0
